Includes packages in the workspaces section while building the component tree
