+++ lib/cobra_commander.rb
@@ -2,6 +2,8 @@
 
 require "cobra_commander/cli"
 require "cobra_commander/component_tree"
+require "cobra_commander/ruby"
+require "cobra_commander/js"
 require "cobra_commander/version"
 require "cobra_commander/graph"
 require "cobra_commander/change"
