+++ spec/fixtures/app/components/h/package.json
@@ -1,13 +1,10 @@
 {
-  "name": "AppName",
+  "name": "h",
   "version": "0.0.1",
   "description": "Some description",
   "author": "Garett Arrowood",
   "license": "UNLICENSED",
-  "workspaces": [
-    "components/h"
-  ],
   "dependencies": {
-    "node_manifest": "link:node_manifest"
+    "f": "link:../f"
   }
 }
